A variety of changes attempting to fix issue number #8 as well as making code improvements and improvements on the logout process.
Even before this commit, the logout process was not functional without using
the view - based login. Also there was a conflict between the redirect directive
used in this package (target) and the default django redirect directive (next).
This was the core of issue number #8.

As its main change, this commit abstracts the LOGIN_URL, which is a native
django setting, from the SHIBBOLETH_LOGIN_URL. If you are using the view-based login,

LOGIN_URL should now point to /shib/login , where the login-view is based, and
SHIBBOLETH_LOGIN_URL shoudl be set to the shibboleth login handler.
https:// example.com/Shibboleth.sso/Login.

You can still avoid the usage of views by pointing LOGIN_URL directly to the
shibboleth login handler. However, this is not recommended.