Fix bug on first-request logins
If the user directly accesses the shibboleth login path (session_key is
None) on Django version < 1.10, in auth.login the session will be persisted
in the database and get deleted again immediately.
This causes the insertion of the slo session mapping to fail (foreign
key constraint).
As a workaround we give him a session to delete.